+++ internal/reconciler/telemetry/log_components_checker.go
@@ -10,7 +10,6 @@ import (
 
 	"github.com/kyma-project/telemetry-manager/apis/telemetry/v1alpha1"
 	"github.com/kyma-project/telemetry-manager/internal/conditions"
-	"github.com/kyma-project/telemetry-manager/internal/slicesext"
 	"strings"
 )
 
@@ -77,14 +76,14 @@ func (l *logComponentsChecker) createMessageForReason(pipelines []v1alpha1.LogPi
 	separator := ","
 	var affectedResources []string
 	if len(pipelines) > 0 {
-		pipelineNames := slicesext.TransformFunc(pipelines, func(p v1alpha1.LogPipeline) string {
+		pipelineNames := extslices.TransformFunc(pipelines, func(p v1alpha1.LogPipeline) string {
 			return p.Name
 		})
 		slices.Sort(pipelineNames)
 		affectedResources = append(affectedResources, fmt.Sprintf("LogPipelines: (%s)", strings.Join(pipelineNames, separator)))
 	}
 	if len(parsers) > 0 {
-		parserNames := slicesext.TransformFunc(parsers, func(p v1alpha1.LogParser) string {
+		parserNames := extslices.TransformFunc(parsers, func(p v1alpha1.LogParser) string {
 			return p.Name
 		})
 		slices.Sort(parserNames)
